feat: 맞춤 공고 스켈레톤 적용
