--- src/page-layout/NoticesLayout/components/AllNoticesSkeleton/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AllNoticesSkeleton";
