+++ src/page-layout/NoticesLayout/components/NoticesSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NoticesSkeleton";
